consts配下にshowReposを切り出す
